+++ gothemis/session/session_test.go
@@ -223,7 +223,29 @@ func testSession(keytype int, t *testing.T) {
 	<-finCh
 }
 
-func testInvalidKeyHandling(t *testing.T) {
+func TestValidKeyHandling(t *testing.T) {
+	kpa, err := keys.New(keys.TypeEC)
+	if nil != err {
+		t.Error(err)
+		return
+	}
+
+	kpb, err := keys.New(keys.TypeEC)
+	if nil != err {
+		t.Error(err)
+		return
+	}
+
+	clb := &testCallbacks{kpa, kpb}
+
+	_, err = New(clientID, kpa.Private, clb)
+	if nil != err {
+		t.Error("Creating Secure Session with EC key")
+		return
+	}
+}
+
+func TestInvalidKeyHandling(t *testing.T) {
 	kpa, err := keys.New(keys.TypeRSA)
 	if nil != err {
 		t.Error(err)
@@ -247,5 +269,4 @@ func testInvalidKeyHandling(t *testing.T) {
 
 func TestSession(t *testing.T) {
 	testSession(keys.TypeEC, t)
-	testInvalidKeyHandling(t)
 }